% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nw_0.R
\name{f_nw_0}
\alias{f_nw_0}
\title{Nadaraya-Watson nonparametric regression.}
\usage{
f_nw_0(x, y, x0, h, k1 = dnorm, ...)
}
\arguments{
\item{x}{vector with values for the regressor variable;}

\item{y}{vector with values for the dependent variable;}

\item{x0}{scalar where the user wants to evaluate the estimator;}

\item{h}{bandwidth (scalar);}

\item{k1}{a kernel function;}

\item{...}{specific parameters for k1.}
}
\value{
scalar which is the Nadaraya-Watson estimate at x0.
}
\description{
Nadaraya-Watson nonparametric regression.
}
\examples{
n <- 300
x <- runif(n, -1, 1)
e <- rnorm(n, 0, 0.5)
y <- sin(0.5 * pi * x) + e
f_nw_0(x = x, y = y, x0 = 0.5, h = 0.1)
}
