% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nw_xg.R
\name{f_nw_grid}
\alias{f_nw_grid}
\title{Nadaraya-Watson nonparametric regression over a grid.}
\usage{
f_nw_grid(x, y, xg, h, k1 = dnorm, ...)
}
\arguments{
\item{x}{vector with values for the regressor variable;}

\item{y}{vector with values for the dependent variable;}

\item{xg}{vector where the user wants to evaluate the estimator;}

\item{h}{bandwidth (scalar);}

\item{k1}{a kernel function;}

\item{...}{specific parameters for k1.}
}
\value{
vector with the Nadaraya-Watson estimates at xg.
}
\description{
Nadaraya-Watson nonparametric regression over a grid.
}
\examples{
n <- 300
x <- runif(n, -1, 1)
e <- rnorm(n, 0, 0.5)
y <- sin(0.5 * pi * x) + e
f_nw_grid(x = x, y = y, xg = x, h = 0.1)
}
