\name{StartupMessage}
\alias{StartupMessage}
\alias{npregStartupMessage}

\title{Startup Message for npreg}

\description{
Prints the startup message when npreg is loaded. Not intended to be called by the user.
}

\details{
The `npreg' ascii start-up message was created using the taag software.
}

\references{
https://patorjk.com/software/taag/
}