\name{npreg-internals}
\alias{npreg-internals}

\title{Internal Functions for "npreg"}

%% List of Internal Functions Called
\alias{bin_sample}
\alias{build_depe}
\alias{build_depe2}
\alias{build_rkhs}
\alias{charkro}
\alias{check_control}
\alias{check_family}
\alias{check_knot}
\alias{check_type}
\alias{check_type2}
\alias{df2lambda}
\alias{fit_gsm}
\alias{fit_sm}
\alias{fit_ssi}
\alias{knot1samp}
\alias{linkinvd}
\alias{pred_depe}
\alias{pred_rkhs}
\alias{print.gsm}
\alias{print.boot.gsm}
\alias{print.sm}
\alias{print.boot.sm}
\alias{print.ss}
\alias{print.boot.ss}
\alias{rexpfam}
\alias{rowkro}
\alias{theta.grad}
\alias{theta.info}
\alias{tune.deep.sm}
\alias{tune.deep.gsm}
\alias{tune.gsm}
\alias{tune.acv.ss}
\alias{tune.aic.ss}
\alias{tune.gacv.ss}
\alias{tune.gcv.ss}
\alias{tune.ocv.ss}
\alias{tune.mle.ss}

\alias{ordkern}
\alias{linkern0}
\alias{linkern1}
\alias{cubkern0}
\alias{cubkern1}
\alias{cubkern2}
\alias{quikern0}
\alias{quikern1}
\alias{quikern2}
\alias{.linkern0}
\alias{.linkern1}
\alias{.cubkern0}
\alias{.cubkern1}
\alias{.cubkern2}
\alias{.quikern0}
\alias{.quikern1}
\alias{.quikern2}
\alias{q2fun}
\alias{q4fun}
\alias{q6fun}

\description{
  Internal functions for "npreg" package.
}
\details{
  These functions are not necessarily intended to be called by the user.
}
\keyword{ internal }